requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.36
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.2
slack_bolt==1.19.1
slack_sdk==3.33.4
alembic==1.13.2
pytest==8.3.3
